chore(*): rename package & extract duplicates to variable
