+++ cli/cmd/root.go
@@ -3,19 +3,15 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/dodopizza/kubectl-shovel/internal/version"
-)
-
-const (
-	dumperImageName = "dodopizza/kubectl-shovel-dumper"
+	"github.com/dodopizza/kubectl-shovel/internal/globals"
 )
 
 // NewShovelCommand initialize Shovel command
 func NewShovelCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:               "kubectl-shovel",
+		Use:               globals.PluginName,
 		Short:             "Get diagnostics from running in k8s dotnet application",
-		Version:           version.GetVersion(),
+		Version:           globals.GetVersion(),
 		SilenceUsage:      true,
 		DisableAutoGenTag: true,
 	}
