+++ internal/globals/version_test.go
@@ -1,4 +1,4 @@
-package version
+package globals
 
 import (
 	"testing"
